MVC design pattern set
array, checking func were separated
